Add behaviour for web client
